+++ src/hooks.ts
@@ -7,6 +7,7 @@ import { checkForHuntPreyGM, checkForHuntPreyPlayer } from "./actions/huntprey.t
 import { targetTokensUnderTemplate, deleteTemplateTargets } from "./templatetarget.ts";
 import { checkForUnstableCheck } from "./effects/unstablecheck.ts";
 import { ChatMessagePF2e, MeasuredTemplateDocumentPF2e } from "foundry-pf2e";
+import { checkIfTemplatePlacedIsStormSpiral } from "./actions/stormspiral.ts";
 
 Hooks.on("init", () => {
     registerSettings();
@@ -19,6 +20,8 @@ Hooks.on('renderChatMessage', async (message: ChatMessagePF2e, html: JQuery<HTML
 Hooks.on("createMeasuredTemplate", async (template: MeasuredTemplateDocumentPF2e, _context, userId) => {
     hook(targetTokensUnderTemplate, template, userId)
             .ifEnabled(SETTINGS.TEMPLATE_TARGET)
+            .run();
+    hook(checkIfTemplatePlacedIsStormSpiral, template)
             .run();
 });