+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
  - Added Storm Spiral Animation macro - no sounds yet
+ - Migrated Storm Spiral from Macro to Action triggered from template hook
 
 ## [0.52.0] - 2025-01-03